// spring-ai-alibaba-studio/spring-ai-alibaba-studio-server/frontend/packages/spark-flow/src/demos/flow-demo/types/flow.ts
export interface IRetryConfig {
  /* Whether to enable retry */
  retry_enabled: boolean;
  /* Maximum number of retries */
  max_retries: number;
  /* Retry interval */
  retry_interval: number;
}

export interface ITryCatchConfig {
  /* Exception handling strategy, defaultValue: default value, failBranch: failure branch, noop: do nothing */
  strategy: 'defaultValue' | 'failBranch' | 'noop';
  /* If strategy is defaultValue, default_value needs to be defined */
  default_values?: Omit<INodeDataInputParamItem, 'value_from'>[];
}
// ...
export interface IApiNodeParam {
  /* Output type, json: output json, consumer uses json_params field; primitive: output raw data, consumer uses result field */
  output_type: 'json' | 'primitive';
  /* Request URL */
  url: string;
  authorization: {
    auth_type: 'NoAuth' | 'ApiKeyAuth' | 'BearerAuth';
    /* When auth_type is NoAuth, auth_config is invalid
    When auth_type is ApiKeyAuth, the key of auth_config can be customized
    When auth_type is BearerAuth, the key of auth_config must be token
    */
    auth_config?: Record<string, string>;
  };
  headers: Array<INodeDataInputParamItem>;
  method: 'post' | 'get' | 'put' | 'delete';
  params: Array<INodeDataInputParamItem>;
  body: {
    type: 'raw' | 'form-data' | 'json' | 'none';
    data?: string | Array<INodeDataInputParamItem>;
  };
  timeout: {
    /* Read timeout */
    read: number;
  };
  /* Retry configuration */
  retry_config: IRetryConfig;
  /* Exception handling configuration */
  try_catch_config: ITryCatchConfig;
}
// ...
export type INodeDataNodeParam =
  | IStartNodeParam
  | IInputNodeParam
  | IEndNodeParam
  | IOutputNodeParam
  | ILLMNodeParam
  | IParameterExtractorNodeParam
  | IJudgeNodeParam
  | IApiNodeParam
  | IParallelNodeParam
  | IIteratorNodeParam
  | IAppComponentNodeParam
  | IPluginNodeParam
  | IClassifierNodeParam
  | IScriptNodeParam
  | IVariableAssignNodeParam
  | IVariableHandleNodeParam
  | IRetrievalNodeParam
  | IMCPNodeParam;

export type IWorkFlowNodeData<T = INodeDataNodeParam> = {
  /* Node name */
  label: string;
  /* Node input parameters */
  input_params: INodeDataInputParamItem[];
  /* Node output parameters */
  output_params: INodeDataOutputParamItem[];
  /* Node business parameters */
  node_param: T;
  /* Node description */
  desc?: string;
};
// ...
export type IApiNodeData = IWorkFlowNodeData<IApiNodeParam>;